refactor: Extract OwnerRow and ReviewRow
