+++ src/components/new-safe/OwnerRow/styles.module.css
@@ -1,0 +1,11 @@
+.name :global .MuiFormHelperText-root {
+  position: absolute;
+  bottom: -20px;
+}
+
+@media (max-width: 900px) {
+  .name :global .MuiFormHelperText-root {
+    position: relative;
+    bottom: 0;
+  }
+}
